refactor(components): Update import paths for ProjectItem type
This commit refactors the import statements in the Projects and ScreenshotCarousel components to source the ProjectItem type from the correct location in the types directory. This change enhances code organization and clarity by separating data types from data sources.
